Add the following: - Educator class: parent class of Teacher and ApprenticeTeacher, move repetitive methods and attributes to parent, add lookup operator for subclass constants as attribute values - Teacher, ApprenticeTeacher classes: make subclass of Educator, add constants - Rspec: pass all rspec tests NOTE: HighFivable module, 'undef_method' to raise error: leftover from original work done during checkpoint challenge LEARNING OUTCOMES: recognize when a separate parent class should be made, solidify constant lookup using lookup operator, solidify requiring modules and class files for rspec tests or for class inheritance or module access, solidify how to exclude a child class from calling on a method/attribute in the parent class
