+++ educator.rb
@@ -1,0 +1,43 @@
+require_relative 'high_fivable'
+
+class Educator
+  include HighFivable
+
+  attr_reader :age, :salary, :phase, :target_raise, :performance_rating
+  attr_accessor :name
+
+  def initialize(options={})
+    @age = options.fetch(:age, 0)
+    @name = options.fetch(:name, "")
+    @phase = 3
+    @salary = 0
+    @target_raise = self.class::TARGET_RAISE
+    @performance_rating = 0
+  end
+
+  def set_phase(num)
+    @phase = num
+    "Cool, I've always wanted to teach phase #{num}!"
+  end
+
+  def set_performance_rating(rating)
+    response = ""
+    if rating > self.class::PERFORMANCE_RATING_THRESHOLD
+      response = "Yay, I'm a great employee!"
+      receive_raise(@target_raise)
+    else
+      response = "Oh, well -- thanks to this actionable, specific, and kind feedback, I'll do better next time."
+    end
+    response
+  end
+
+  def receive_raise(raise)
+    @salary += raise
+  end
+
+  def salary=(new_salary)
+    puts "This better be good!"
+    @salary = new_salary
+  end
+
+end
