+++ student.rb
@@ -1,4 +1,7 @@
+require_relative 'high_fivable'
+
 class Student
+  include HighFivable
   attr_reader :age, :phase
   attr_accessor :name
 
@@ -6,10 +9,6 @@ class Student
     @phase = 1
     @age = options.fetch(:age, 0)
     @name = options.fetch(:name, "")
-  end
-
-  def offer_high_five
-    "High five!"
   end
 
   def set_phase(num)
